fix: E2E 테스트 피벗 반영 — vote/result 기준 재작성
- battle/feed E2E 삭제 → vote.spec (질문+캐릭터+투표+군중) + result.spec + 탭 전환
- 3개 E2E 통과

Closes #37
